backend/requirements.txt
Django==5.2.3
pymongo==4.6.1
djangorestframework==3.14.0
